chore: split fixtures to own packages
