+++ fixtures/expansion-panel/src/expansion-panel.ts
