+++ fixtures/intl-currency/src/intl-currency.ts
