+++ fixtures/progress-bar/src/progress-bar.ts
